feat: Add onionspray_project_settings variable
This variable now contains all values used to customize a project, and
is a list of dict. To handle more than one project on the same
Onionspray deployment, we can now add a new dict to this list.

We split the handler of the config into two, so that we can loop through
the onionspray_project_settings variable more easily.