feat: Add variables and checks to handle CA cert verifications
Two new variables are introduced, the first one to specify a custom path
for a CA file, and the second one to enable or disable checking the
upstream certificate against the specified CA(s). The role now also
checks whether the CA file exists, and aborts if it does not.